move remaining components out of @nteract/core
