+++ applications/showcase/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@nteract/commutable": "^4.0.1",
+    "@nteract/connected-components": "^1.0.0",
     "@nteract/core": "^5.0.1",
     "@nteract/display-area": "^4.1.1",
     "@nteract/dropdown-menu": "^0.1.2",
