+++ packages/connected-components/src/index.js
