+++ packages/connected-components/src/title-bar/logos.js
