+++ packages/core/package.json
@@ -19,9 +19,6 @@
     "@nteract/messaging": "^4.0.0",
     "escape-carriage": "^1.2.0",
     "file-saver": "^1.3.3",
-    "prop-types": "^15.6.1",
-    "rc-menu": "6.2.10",
-    "react-redux": "^5.0.7",
     "redux": "^4.0.0",
     "redux-immutable": "^4.0.0",
     "redux-observable": "^0.18.0",
@@ -33,17 +30,14 @@
     "uuid": "^3.1.0"
   },
   "peerDependencies": {
-    "immutable": "^4.0.0-rc.9",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "styled-jsx": "^2.2.6"
+    "immutable": "^4.0.0-rc.9"
   },
   "publishConfig": {
     "access": "public"
   },
   "keywords": [
     "nteract",
-    "ohgodamonomoduleinamonorepo"
+    "redux"
   ],
   "author": "Kyle Kelley <rgbkrk@gmail.com>",
   "license": "BSD-3-Clause"
