+++ packages/timeago/src/language-strings/ar.js
@@ -9,7 +9,11 @@ function numpf(n, a) {
         ? 1
         : n === 2
           ? 2
-          : n % 100 >= 3 && n % 100 <= 10 ? 3 : n % 100 >= 11 ? 4 : 5
+          : n % 100 >= 3 && n % 100 <= 10
+            ? 3
+            : n % 100 >= 11
+              ? 4
+              : 5
   ];
 }
