+++ packages/timeago/src/language-strings/sl.js
@@ -6,7 +6,11 @@ function numpf(n, a) {
   return a[
     n % 100 === 1
       ? 1
-      : n % 100 === 2 ? 2 : n % 100 === 3 || n % 100 === 4 ? 3 : 0
+      : n % 100 === 2
+        ? 2
+        : n % 100 === 3 || n % 100 === 4
+          ? 3
+          : 0
   ];
 }
